Otimização do Gerenciador de Tarefas

+++ src/pages/projeto/projeto.ts
@@ -1,7 +1,8 @@
 import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams, ToastController } from 'ionic-angular';
+import { IonicPage, NavController, NavParams, ToastController, Platform } from 'ionic-angular';
 import { ProjetosServiceProvider } from '../../providers/projetos-service/projetos-service';
-/*teste*/
+import { Dialogs } from '@ionic-native/dialogs';
+
 @IonicPage()
 @Component({
   selector: 'page-projeto',
@@ -17,7 +18,9 @@ export class ProjetoPage {
   constructor(public navCtrl: NavController, 
     public navParams: NavParams,
     public projetosService: ProjetosServiceProvider,
-    public toastCtrl: ToastController) {
+    public toastCtrl: ToastController,
+    private dialogs: Dialogs,
+    public platform: Platform) {
       this.codigoProjeto = navParams.get('codigo');
       this.novo = navParams.get('novo');
 
@@ -40,26 +43,55 @@ export class ProjetoPage {
   }
 
   alterar(){
-    this.projetosService.editProjetos(this.codigoProjeto,this.nomeProjeto)
-    .then(dados => {
-      this.presentToast('Projeto alterado com sucesso!');
-      this.navCtrl.pop();
+    this.platform.ready().then(()=> {
+      this.dialogs.alert('O projeto sera alterado?',
+     'Alerta de Alteração','OK')
+     .then((dados) => {
+      if(dados === 1){
+        this.projetosService.editProjetos(this.codigoProjeto,this.nomeProjeto)
+        .then(dados => {
+          this.presentToast('Projeto alterado com sucesso!');
+          this.navCtrl.pop();
+        });
+      }
+     })
+     .catch( e => console.log('Error displaying dialog', e));
     });
+    
   }
 
   excluir(){
-    this.projetosService.deleteProjetos(this.codigoProjeto)
-    .then(dados => {
-      this.navCtrl.pop();
-    });
+
+    this.platform.ready().then(() => {
+      this.dialogs.confirm('Deseja realmente excluir este projeto.',
+     'Confirmação',['Ok','Cancel'] )
+     .then((dados) => {
+       if(dados === 1){
+        this.projetosService.deleteProjetos(this.codigoProjeto)
+        .then(dados => {
+          this.navCtrl.pop();
+        });
+       }
+    })
+    .catch( e => console.log('Error displaying dialog', e));
+    });     
   }
 
   incluir(){
-    this.projetosService.addProjeto(this.nomeProjeto)
-    .then(dados =>{
-      this.presentToast('Projeto adicionado com sucesso!');
-      this.navCtrl.pop();
-    });
+    this.platform.ready().then(() => {
+       this.dialogs.confirm('Deseja adiconar este projeto.',
+      'Confirmação',['Ok','Cancel'] )
+      .then((dados) => {
+        if(dados === 1 ){
+          this.projetosService.addProjeto(this.nomeProjeto)
+          .then(dados =>{
+            this.presentToast('Projeto adicionado com sucesso!');  
+            //this.onConfirm();    
+            this.navCtrl.pop();
+          });
+        }
+      })// console.log('Dialog dismissed'))
+      .catch( e => console.log('Error displaying dialog', e));
+    });    
   }
-
 }
